fix(frontend): fix mutation exports, remove unused attributes/imports
