+++ app/client/src/features/events/LiveFeedback/LiveFeedbackDialog.tsx
@@ -1,8 +1,9 @@
 import * as React from 'react';
 import { DialogContent } from '@material-ui/core';
-import { useMutation, graphql } from 'react-relay';
+import { useMutation } from 'react-relay';
 
 import type { SubmitLiveFeedbackMutation } from '@local/__generated__/SubmitLiveFeedbackMutation.graphql';
+import { SUBMIT_LIVE_FEEDBACK_MUTATION } from './SubmitLiveFeedback'
 import { ResponsiveDialog } from '@local/components/ResponsiveDialog';
 import { LiveFeedbackForm, TLiveFeedbackFormState } from './LiveFeedbackForm';
 
@@ -13,28 +14,6 @@ interface Props {
     close: () => void;
     eventId: string;
 }
-
-export const SUBMIT_LIVE_FEEDBACK_MUTATION = graphql`
-    mutation SubmitLiveFeedbackMutation($input: CreateFeedback!) {
-        createFeedback(input: $input) {
-            isError
-            message
-            body {
-                cursor
-                node {
-                    id
-                    createdAt
-                    message
-                    createdBy {
-                        id
-                        firstName
-                        lastName
-                    }
-                }
-            }
-        }
-    }
-`;
 
 export function LiveFeedbackDialog({ isOpen, openLinked, close, eventId }: Props) {
     const [commit] = useMutation<SubmitLiveFeedbackMutation>(SUBMIT_LIVE_FEEDBACK_MUTATION);
@@ -49,7 +28,6 @@ export function LiveFeedbackDialog({ isOpen, openLinked, close, eventId }: Props
     return (
         <ResponsiveDialog
             open={isOpen}
-            currDialog='feedback'
             onClose={close}
         >
             <DialogContent>
